Remove createSolValueJSX, replace on DisplayValue everywhere
